requirements-dev.txt
python-dotenv
pydantic
snowflake-connector-python[pandas]
snowflake-sqlalchemy
sqlalchemy
psycopg2-binary
pandas
numpy
pyarrow
great-expectations
pyspark
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
dash
dash-bootstrap-components
plotly
loguru
tenacity
rich
typer